requirements.txt
flask>=3.0
flask-cors>=4.0
googlemaps>=4.10
ortools>=9.10
gunicorn>=21.0
flask-sqlalchemy>=3.0
psycopg2-binary>=2.9
fpdf2>=2.7
sendgrid>=6.10
